requirements.txt
aerich==0.9.1
aiosmtplib==3.0.2
aiosqlite==0.21.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
asyncclick==8.2.2.2
asyncpg==0.30.0
atlastk==0.13.5
bcrypt==4.3.0
black==25.1.0
blinker==1.9.0
cffi==1.17.1
cfgv==3.4.0
click==8.2.1
cryptography==45.0.6
dictdiffer==0.9.0
distlib==0.4.0
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.2.0
fastapi==0.116.1
fastapi-mail==1.5.0
fastapi-users==14.0.1
fastapi-users-db-sqlalchemy==7.0.0
fastapi-users-tortoise==0.2.0
filelock==3.19.1
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
identify==2.6.13
idna==3.10
iso8601==2.1.0
jinja2==3.1.6
makefun==1.16.0
mako==1.3.10
markupsafe==3.0.2
mypy-extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.8
pre-commit==4.3.0
pwdlib==0.2.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pyjwt==2.10.1
pypika-tortoise==0.6.1
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
rsa==4.9.1
ruff==0.12.10
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.43
starlette==0.47.3
tortoise==0.1.1
tortoise-orm==0.25.1
typing-extensions==4.15.0
typing-inspection==0.4.1
uv==0.8.13
uvicorn==0.35.0
uvloop==0.21.0
virtualenv==20.34.0
watchfiles==1.1.0
websockets==15.0.1

pydantic[email]